Revert "Add (buggy) kmalloc, refactor some stuff, make klog take varargs like printf, try to work out how to do memory detection properly with limine (not going well)"
This reverts commit e39488c7d0d6c79c7aa0d2898ac70d55939ae526.

+++ kernel/src/acpi/acpi.c
@@ -1,4 +1,3 @@
-#include <stdint.h>
 #include <string.h>
 #include <term/term.h>
 #include <klog/klog.h>
@@ -11,9 +10,7 @@ void acpi_init(XSDP_t* xsdp_table)
     {
         klog("acpi", "Assuming ACPI 1.0");
         term_printf("ACPI Revision: %d\n", xsdp_table->revision);
-        // pretty sure this next line is anything but safe... casting a 32 bit
-        // address to 64 bit and then expecting it to "just work" as a pointer
-        rsdt = (RSDT_t *) ((uint64_t)xsdp_table->RSDT_address);
+        rsdt = (RSDT_t *) (xsdp_table->RSDT_address);
         if (memcmp(xsdp_table->signature, "RSD PTR ", 8) != 0)
         {
             klog("acpi", "ACPI signature match error");
